Move registry to a new module
